Created main.py
This file serves as a tutorial for functions in pygame - as well as casual fun.
Disclaimer: Requires Python and pygame to run. You will also require the sounds to run the game. If you don't want sounds, just add a hashtag # at the start of the lines that use the pygame.mixer modul